+++ src/Page/Dashboard/DashboardMenu.tsx
@@ -17,7 +17,7 @@ import {
   DASHBOARD_URL,
   ICON_NAMES,
   MAINNET_NETWORK,
-  NEW_ACCOUNT_ADD_TAB_URL,
+  NEW_ACCOUNT_ADD_URL,
   NEW_ACCOUNT_SUB_URL,
   NEW_ACCOUNT_IMPORT_URL,
   REMOVE_ACCOUNT_URL,
@@ -203,9 +203,9 @@ export const DashboardMenu: React.FC = () => {
           </WalletCopy>
           {menuTargetAccountLevel !== 'addressIndex' && (
             <WalletAction
-              onClick={() => navigate(NEW_ACCOUNT_SUB_URL)}
+              onClick={() => openTab(NEW_ACCOUNT_SUB_URL)}
               tabIndex={0}
-              onKeyPress={(e) => keyPress(e, () => navigate(NEW_ACCOUNT_SUB_URL))}
+              onKeyPress={(e) => keyPress(e, () => openTab(NEW_ACCOUNT_IMPORT_URL))}
             >
               Create Sub Account
             </WalletAction>
@@ -245,12 +245,12 @@ export const DashboardMenu: React.FC = () => {
       )}
       <BottomFloat>
         <ButtonGroup>
-          <Button onClick={() => openTab(NEW_ACCOUNT_ADD_TAB_URL)}>
+          <Button onClick={() => openTab(NEW_ACCOUNT_ADD_URL)}>
             Create New Account
           </Button>
           <Button
             variant="secondary"
-            onClick={() => navigate(NEW_ACCOUNT_IMPORT_URL)}
+            onClick={() => openTab(NEW_ACCOUNT_IMPORT_URL)}
           >
             Import Account
           </Button>
