+++ src/Page/Landing/Landing.tsx
@@ -1,11 +1,7 @@
 import { BottomFloat, Button, ButtonGroup, Content } from 'Components';
 import { useNavigate } from 'react-router-dom';
 import { Carousel } from './Carousel';
-import {
-  NEW_ACCOUNT_CREATE_TAB_URL,
-  NEW_ACCOUNT_RECOVER_URL,
-  UNLOCK_URL,
-} from 'consts';
+import { NEW_ACCOUNT_CREATE_URL, NEW_ACCOUNT_RECOVER_URL, UNLOCK_URL } from 'consts';
 import { useActiveAccount } from 'redux/hooks';
 import { openTab } from 'utils';
 
@@ -16,9 +12,7 @@ export const Landing: React.FC = () => {
     masterKey ? (
       <Button onClick={() => navigate(UNLOCK_URL)}>Unlock</Button>
     ) : (
-      <Button onClick={() => openTab(NEW_ACCOUNT_CREATE_TAB_URL)}>
-        Create Wallet
-      </Button>
+      <Button onClick={() => openTab(NEW_ACCOUNT_CREATE_URL)}>Create Wallet</Button>
     );
 
   return (
@@ -29,7 +23,7 @@ export const Landing: React.FC = () => {
           {renderLandingActions()}
           <Button
             variant="transparent"
-            onClick={() => navigate(NEW_ACCOUNT_RECOVER_URL)}
+            onClick={() => openTab(NEW_ACCOUNT_RECOVER_URL)}
           >
             Recover Wallet
           </Button>
