+++ src/Page/NewAccount/index.ts
@@ -1,5 +1,7 @@
-export * from './SeedphraseInfo';
-export * from './SeedphraseInput';
+export * from './NewAccountName';
 export * from './NewAccountPassword';
 export * from './NewAccountSuccess';
-export * from './tab';
+export * from './SeedphraseInfo';
+export * from './SeedphraseInput';
+export * from './SeedphraseValue';
+export * from './SeedphraseVerify';
