--- src/Page/NewAccount/tab/index.ts
@@ -1,6 +1,0 @@
-export * from './NewAccountNameTab';
-export * from './SeedphraseInfoTab';
-export * from './SeedphraseValueTab';
-export * from './SeedphraseVerifyTab';
-export * from './NewAccountPasswordTab';
-export * from './NewAccountSuccessTab';
