+++ src/consts/urls.ts
@@ -15,44 +15,6 @@ export const EXPLORER_WEB_BLOCK = `/block`;
 // ---------------------
 // ROOT HOMEPAGE URL
 export const APP_URL = '/';
-// NEW ACCOUNT (ADD/CREATE/RECOVER)
-export const NEW_ACCOUNT_URL = '/account';
-const NEW_ACCOUNT_SEED_INFO = '/seed';
-const NEW_ACCOUNT_SEED_VALUE = `${NEW_ACCOUNT_SEED_INFO}/value`;
-const NEW_ACCOUNT_SEED_VERIFY = `${NEW_ACCOUNT_SEED_INFO}/verify`;
-const NEW_ACCOUNT_SEED_INPUT = `${NEW_ACCOUNT_SEED_INFO}/input`;
-const NEW_ACCOUNT_PASSWORD = '/password';
-const NEW_ACCOUNT_SUCCESS = '/success';
-// - SUB ACCOUNT
-export const NEW_ACCOUNT_SUB_URL = `${NEW_ACCOUNT_URL}/sub`;
-export const SUB_ACCOUNT_PASSWORD_URL = `${NEW_ACCOUNT_SUB_URL}${NEW_ACCOUNT_PASSWORD}`;
-export const SUB_ACCOUNT_SUCCESS_URL = `${NEW_ACCOUNT_SUB_URL}${NEW_ACCOUNT_SUCCESS}`;
-// - CREATE
-export const NEW_ACCOUNT_CREATE_URL = `${NEW_ACCOUNT_URL}/create`;
-export const CREATE_ACCOUNT_SEED_INFO_URL = `${NEW_ACCOUNT_CREATE_URL}${NEW_ACCOUNT_SEED_INFO}`;
-export const CREATE_ACCOUNT_SEED_VALUE_URL = `${NEW_ACCOUNT_CREATE_URL}${NEW_ACCOUNT_SEED_VALUE}`;
-export const CREATE_ACCOUNT_SEED_VERIFY_URL = `${NEW_ACCOUNT_CREATE_URL}${NEW_ACCOUNT_SEED_VERIFY}`;
-export const CREATE_ACCOUNT_PASSWORD_URL = `${NEW_ACCOUNT_CREATE_URL}${NEW_ACCOUNT_PASSWORD}`;
-export const CREATE_ACCOUNT_SUCCESS_URL = `${NEW_ACCOUNT_CREATE_URL}${NEW_ACCOUNT_SUCCESS}`;
-// - RECOVER
-export const NEW_ACCOUNT_RECOVER_URL = `${NEW_ACCOUNT_URL}/recover`;
-export const RECOVER_ACCOUNT_SEED_INFO_URL = `${NEW_ACCOUNT_RECOVER_URL}${NEW_ACCOUNT_SEED_INFO}`;
-export const RECOVER_ACCOUNT_SEED_INPUT_URL = `${NEW_ACCOUNT_RECOVER_URL}${NEW_ACCOUNT_SEED_INPUT}`;
-export const RECOVER_ACCOUNT_PASSWORD_URL = `${NEW_ACCOUNT_RECOVER_URL}${NEW_ACCOUNT_PASSWORD}`;
-export const RECOVER_ACCOUNT_SUCCESS_URL = `${NEW_ACCOUNT_RECOVER_URL}${NEW_ACCOUNT_SUCCESS}`;
-// - ADD
-export const NEW_ACCOUNT_ADD_URL = `${NEW_ACCOUNT_URL}/add`;
-export const ADD_ACCOUNT_SEED_INFO_URL = `${NEW_ACCOUNT_ADD_URL}${NEW_ACCOUNT_SEED_INFO}`;
-export const ADD_ACCOUNT_SEED_VALUE_URL = `${NEW_ACCOUNT_ADD_URL}${NEW_ACCOUNT_SEED_VALUE}`;
-export const ADD_ACCOUNT_SEED_VERIFY_URL = `${NEW_ACCOUNT_ADD_URL}${NEW_ACCOUNT_SEED_VERIFY}`;
-export const ADD_ACCOUNT_PASSWORD_URL = `${NEW_ACCOUNT_ADD_URL}${NEW_ACCOUNT_PASSWORD}`;
-export const ADD_ACCOUNT_SUCCESS_URL = `${NEW_ACCOUNT_ADD_URL}${NEW_ACCOUNT_SUCCESS}`;
-// - IMPORT
-export const NEW_ACCOUNT_IMPORT_URL = `${NEW_ACCOUNT_URL}/import`;
-export const IMPORT_ACCOUNT_SEED_INFO_URL = `${NEW_ACCOUNT_IMPORT_URL}${NEW_ACCOUNT_SEED_INFO}`;
-export const IMPORT_ACCOUNT_SEED_INPUT_URL = `${NEW_ACCOUNT_IMPORT_URL}${NEW_ACCOUNT_SEED_INPUT}`;
-export const IMPORT_ACCOUNT_PASSWORD_URL = `${NEW_ACCOUNT_IMPORT_URL}${NEW_ACCOUNT_PASSWORD}`;
-export const IMPORT_ACCOUNT_SUCCESS_URL = `${NEW_ACCOUNT_IMPORT_URL}${NEW_ACCOUNT_SUCCESS}`;
 // DASHBOARD URLS
 export const DASHBOARD_URL = '/dashboard';
 export const DASHBOARD_MENU_URL = `${DASHBOARD_URL}/menu`;
@@ -90,18 +52,41 @@ export const RESET_WALLET_URL = '/reset';
 // INNER-APP TAB URLS
 // ---------------------
 export const NEW_TAB_ROOT_URL = 'index.html';
-// NEW ACCOUNT (ADD/CREATE)
+// NEW ACCOUNT (ADD/CREATE/RECOVER/IMPORT)
+export const NEW_ACCOUNT_URL = '/account';
+const NEW_ACCOUNT_SEED_INFO = '/seed';
+const NEW_ACCOUNT_SEED_VALUE = `${NEW_ACCOUNT_SEED_INFO}/value`;
+const NEW_ACCOUNT_SEED_VERIFY = `${NEW_ACCOUNT_SEED_INFO}/verify`;
+const NEW_ACCOUNT_SEED_INPUT = `${NEW_ACCOUNT_SEED_INFO}/input`;
+const NEW_ACCOUNT_PASSWORD = '/password';
+const NEW_ACCOUNT_SUCCESS = '/success';
 // - CREATE
-export const NEW_ACCOUNT_CREATE_TAB_URL = `#${NEW_ACCOUNT_URL}/create`;
-export const CREATE_ACCOUNT_SEED_INFO_TAB_URL = `${NEW_ACCOUNT_CREATE_TAB_URL}${NEW_ACCOUNT_SEED_INFO}`;
-export const CREATE_ACCOUNT_SEED_VALUE_TAB_URL = `${NEW_ACCOUNT_CREATE_TAB_URL}${NEW_ACCOUNT_SEED_VALUE}`;
-export const CREATE_ACCOUNT_SEED_VERIFY_TAB_URL = `${NEW_ACCOUNT_CREATE_TAB_URL}${NEW_ACCOUNT_SEED_VERIFY}`;
-export const CREATE_ACCOUNT_PASSWORD_TAB_URL = `${NEW_ACCOUNT_CREATE_TAB_URL}${NEW_ACCOUNT_PASSWORD}`;
-export const CREATE_ACCOUNT_SUCCESS_TAB_URL = `${NEW_ACCOUNT_CREATE_TAB_URL}${NEW_ACCOUNT_SUCCESS}`;
+export const NEW_ACCOUNT_CREATE_URL = `#${NEW_ACCOUNT_URL}/create`;
+export const CREATE_ACCOUNT_SEED_INFO_URL = `${NEW_ACCOUNT_CREATE_URL}${NEW_ACCOUNT_SEED_INFO}`;
+export const CREATE_ACCOUNT_SEED_VALUE_URL = `${NEW_ACCOUNT_CREATE_URL}${NEW_ACCOUNT_SEED_VALUE}`;
+export const CREATE_ACCOUNT_SEED_VERIFY_URL = `${NEW_ACCOUNT_CREATE_URL}${NEW_ACCOUNT_SEED_VERIFY}`;
+export const CREATE_ACCOUNT_PASSWORD_URL = `${NEW_ACCOUNT_CREATE_URL}${NEW_ACCOUNT_PASSWORD}`;
+export const CREATE_ACCOUNT_SUCCESS_URL = `${NEW_ACCOUNT_CREATE_URL}${NEW_ACCOUNT_SUCCESS}`;
 // - ADD
-export const NEW_ACCOUNT_ADD_TAB_URL = `#${NEW_ACCOUNT_URL}/add`;
-export const ADD_ACCOUNT_SEED_INFO_TAB_URL = `${NEW_ACCOUNT_ADD_TAB_URL}${NEW_ACCOUNT_SEED_INFO}`;
-export const ADD_ACCOUNT_SEED_VALUE_TAB_URL = `${NEW_ACCOUNT_ADD_TAB_URL}${NEW_ACCOUNT_SEED_VALUE}`;
-export const ADD_ACCOUNT_SEED_VERIFY_TAB_URL = `${NEW_ACCOUNT_ADD_TAB_URL}${NEW_ACCOUNT_SEED_VERIFY}`;
-export const ADD_ACCOUNT_PASSWORD_TAB_URL = `${NEW_ACCOUNT_ADD_TAB_URL}${NEW_ACCOUNT_PASSWORD}`;
-export const ADD_ACCOUNT_SUCCESS_TAB_URL = `${NEW_ACCOUNT_ADD_TAB_URL}${NEW_ACCOUNT_SUCCESS}`;
+export const NEW_ACCOUNT_ADD_URL = `#${NEW_ACCOUNT_URL}/add`;
+export const ADD_ACCOUNT_SEED_INFO_URL = `${NEW_ACCOUNT_ADD_URL}${NEW_ACCOUNT_SEED_INFO}`;
+export const ADD_ACCOUNT_SEED_VALUE_URL = `${NEW_ACCOUNT_ADD_URL}${NEW_ACCOUNT_SEED_VALUE}`;
+export const ADD_ACCOUNT_SEED_VERIFY_URL = `${NEW_ACCOUNT_ADD_URL}${NEW_ACCOUNT_SEED_VERIFY}`;
+export const ADD_ACCOUNT_PASSWORD_URL = `${NEW_ACCOUNT_ADD_URL}${NEW_ACCOUNT_PASSWORD}`;
+export const ADD_ACCOUNT_SUCCESS_URL = `${NEW_ACCOUNT_ADD_URL}${NEW_ACCOUNT_SUCCESS}`;
+// - RECOVER
+export const NEW_ACCOUNT_RECOVER_URL = `#${NEW_ACCOUNT_URL}/recover`;
+export const RECOVER_ACCOUNT_SEED_INFO_URL = `${NEW_ACCOUNT_RECOVER_URL}${NEW_ACCOUNT_SEED_INFO}`;
+export const RECOVER_ACCOUNT_SEED_INPUT_URL = `${NEW_ACCOUNT_RECOVER_URL}${NEW_ACCOUNT_SEED_INPUT}`;
+export const RECOVER_ACCOUNT_PASSWORD_URL = `${NEW_ACCOUNT_RECOVER_URL}${NEW_ACCOUNT_PASSWORD}`;
+export const RECOVER_ACCOUNT_SUCCESS_URL = `${NEW_ACCOUNT_RECOVER_URL}${NEW_ACCOUNT_SUCCESS}`;
+// - SUB ACCOUNT
+export const NEW_ACCOUNT_SUB_URL = `#${NEW_ACCOUNT_URL}/sub`;
+export const SUB_ACCOUNT_PASSWORD_URL = `${NEW_ACCOUNT_SUB_URL}${NEW_ACCOUNT_PASSWORD}`;
+export const SUB_ACCOUNT_SUCCESS_URL = `${NEW_ACCOUNT_SUB_URL}${NEW_ACCOUNT_SUCCESS}`;
+// - IMPORT
+export const NEW_ACCOUNT_IMPORT_URL = `#${NEW_ACCOUNT_URL}/import`;
+export const IMPORT_ACCOUNT_SEED_INFO_URL = `${NEW_ACCOUNT_IMPORT_URL}${NEW_ACCOUNT_SEED_INFO}`;
+export const IMPORT_ACCOUNT_SEED_INPUT_URL = `${NEW_ACCOUNT_IMPORT_URL}${NEW_ACCOUNT_SEED_INPUT}`;
+export const IMPORT_ACCOUNT_PASSWORD_URL = `${NEW_ACCOUNT_IMPORT_URL}${NEW_ACCOUNT_PASSWORD}`;
+export const IMPORT_ACCOUNT_SUCCESS_URL = `${NEW_ACCOUNT_IMPORT_URL}${NEW_ACCOUNT_SUCCESS}`;
